Fix 2 legacy bugs
Bug 1: at first frame after loading, after deleting obj_mapsample (loading screen plane), value obj_mapsample is not set to a default value. Hence when the game is restarted, obj_mapsample will be deleted again but this time it can be another object actually within the engine. Hence we need to do this only if obj_mapsample is not -1.
This caused game unable to restart.

Bug 2: if system runs for a long-long time, GetTickCount() may return value that being stored as an integer, could overflow. If game is restarted, fps_ms will be set to 0, and inside FrameLimiter(), the value of millisecs for sleeping will be calculated as a very big positive number.