Updated debug settings and lowprio bugfix
Added extra debug settings (remote debug and TD32).
Minor bugfix: made sure to display HUD and Xhair only after the sliding/opening planes at the beginning of a new game slided out of the screen. With the new engine, xhair was still visible over these planes, but actually it is a bug of the legacy game that was exposed by the new engine. It was lucky behavior with legacy engine.