Fixed marks on walls (bulletholes)
Since PR00FPS runs with new engine, marks on walls were not visible.
Setting them to doublesided solved the issue, this proved this was just a facing problem. Since we dont want extra rendering due to being doublesided, they stay one sided, but the angle they face is mirrored now so they comply with the x-y-z coordinate system of the new engine.